KAIST_2024/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kaist-diffusion-assignments"
version = "0.1.0"
description = "KAIST 2024 Diffusion Model Assignments"
authors = [
    {name = "Student", email = "student@kaist.ac.kr"}
]
requires-python = ">=3.9"
dependencies = [
    # Common dependencies that might be used across assignments
    "torch",
    "torchvision",
    "numpy>=1.24.1",
    "matplotlib>=3.6.0",
    "pillow",
    "tqdm>=4.64.1",
    "jupyter",
    "jupyterlab>=4.0.2",
    "ipython>=8.12.0",
    # Additional dependencies from assignment requirements
    "scikit-learn>=1.1.3",
    "torch-ema>=0.3",
    "jaxtyping>=0.2.20",
    "pytorch_lightning",
    "dotmap",
    # Assignment 4 dependencies (compatible versions)
    "accelerate>=0.25.0",
    "diffusers>=0.19.3",
    "einops>=0.7.0",
    "transformers>=4.28.1",
    "huggingface-hub>=0.24.5",
    "safetensors>=0.4.4",
    "ftfy>=6.2.3",
    "sentencepiece>=0.1.97",
    # Note: CLIP will be installed separately if needed
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tokenizers>=0.22.1",
    "clip",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["zip*", "*.zip", "__pycache__*", "*.pyc"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git", rev = "dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1" }

[tool.uv.workspace]
members = [
    "Diffusion-Assignment3-ControlNet-LoRA-master",
]
